% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaahoopsscraper.R
\name{ncaahoopsscraper}
\alias{ncaahoopsscraper}
\title{ncaahoopsscraper}
\usage{
ncaahoopsscraper(x, y)
}
\arguments{
\item{x}{The beginning date of data scraping, in the format
"yyyy/mm/dd".}

\item{y}{The ending date of data scraping, in the format
"yyyy/mm/dd".}
}
\value{
A tidy dataframe with the following variables for all games within
the date range: GameDate, HomeTeamName, HomeTeamScore, AwayTeamName,
AwayTeamScore.
}
\description{
Scrapes scores for any season of college basketball.
}
\details{
Takes inputs for beginning and end of season and scrapes www.sports-reference.com for all scores between those dates (inclusive).
}
\examples{
\dontrun{ncaahoopsscraper("2017/11/10", "2018/04/02")}
}
\author{
Kurt Wirth
}
